feat: implemented the actual recursive deletion of files
Functionality is based off of what is in the search hit list in real time as the function iterates through the items in it. It is based off of the folder query file hit list that is present when the delete button is pressed, any file that is moved, deleted, or is otherwise inaccessible when confirmation is given will be counted as an error and not deleted (this solves undesirable behavior that could arise from potential race conditions).